desktop-icons: Improve PDF encryption detection logic
Look in the tail of the pdf for encryption string and old style trailer or XRefStream

If there is an encryption string with a trailer or XRefstream, the file is encrypted.

However the password may not be set and the file may still be Readable! It is encrypted in the pdf specification sense.