Prunes WebViews aggressively
Ensures WebViews are destroyed and detached when no longer needed aggressively

This change introduces an `isAlive` check to avoid operations on destroyed
WebViews and ensures proper disposal of WebViews and associated resources
during widget removal.

The isAlive check is used so that aggressive pruning of webviews does not destroy the webcontext if host is alive but the webview is still not ready as it is async.